Enable the Gst filter only when needed
This is an attempt to work around an issue where the stream will freeze
at the start of songs. I haven't been able to reproduce it on a dev
machine, and it's not certain that it's caused by the plugin. The
downside is a pause as the filter is inserted.